This GitHub repository hosts a Java-based Library Management System. It includes classes for managing library items, such as books, along with features for adding, borrowing, returning, and listing available and checked-out books. The system showcases object-oriented programming concepts and provides a foundation for efficient library resource management.  Explore the code and enhance it to suit your library's specific needs or use it as a learning resource for object-oriented programming in Java. Feel free to contribute and collaborate to make this system even better!
This GitHub repository contains a Library Management System implemented in Java. The system provides a user-friendly interface to manage a library's inventory, featuring functionalities for adding new books, borrowing and returning books, and listing available and checked-out books. It serves as an illustrative example of object-oriented programming in Java, showcasing inheritance, setters, getters, overloading, overriding, and encapsulation.

Whether you're a Java developer looking to explore object-oriented principles or someone interested in managing a personal library, this project offers a foundation to build upon. You're welcome to fork, modify, and collaborate on this project to tailor it to your specific library management needs or to deepen your understanding of Java programming.

Feel free to contribute, report issues, or share your ideas to make this Library Management System even more robust and useful.